Rework CacheDB to SSZCache: no internal `db` object here
