+++ ssz/cache/utils.py
