+++ ssz/sedes/basic.py
@@ -22,7 +22,7 @@ from eth_utils.toolz import (
 from ssz import (
     constants,
 )
-from ssz.cache import (
+from ssz.cache.utils import (
     get_key,
 )
 from ssz.exceptions import (
