+++ tests/misc/test_cache.py
@@ -5,9 +5,8 @@ from ssz import (
     bytes32,
     uint8,
 )
-from tests.utils.db import (
-    CacheDB,
-    MemoryDB,
+from ssz.cache.cache import (
+    SSZCache,
 )
 
 
@@ -23,7 +22,7 @@ def foo_with_db():
     return Foo(
         field1=10,
         field2=b'\x12' * 32,
-        cache=CacheDB(db=MemoryDB(), cache_size=2**10),
+        cache=SSZCache(cache_size=2**10),
     )
